Cambios a extension ".c" Los he puesto en ".h" para que sean librerias
