+++ Ahorcado/principal.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <string.h>
-#include "dibujos_ahorcado.c"
-#include "palabras_ahorcado.c"
+#include "dibujos_ahorcado.h"
+#include "palabras_ahorcado.h"
 int main(int argc, char const *argv[]) {
 
   printf("\n%s\n", Palabra());
